EASM/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.6.3
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
cleanco==2.3
click==8.1.7
click-plugins==1.1.1
colorama==0.4.6
cryptography==45.0.3
distro==1.9.0
dnspython==2.7.0
fake-useragent==2.2.0
filelock==3.18.0
Flask==2.3.3
Flask-Login==0.6.3
flask-talisman==1.1.0
frozenlist==1.6.0
gitdb==4.0.12
GitPython==3.1.44
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.1.2
Jinja2==3.1.2
jiter==0.10.0
jmespath==1.0.1
MarkupSafe==2.1.3
multidict==6.4.4
openai==0.28.1
packaging==25.0
pillow==11.2.1
propcache==0.3.1
psutil==7.0.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pyOpenSSL==25.1.0
PySocks==1.7.1
python-dotenv==1.1.0
pytz==2025.2
RapidFuzz==3.13.0
reportlab==4.4.1
requests==2.32.3
requests-file==2.1.0
schedule==1.2.2
setuptools==78.1.1
shodan==1.31.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
stem==1.8.2
tldextract==5.3.0
toml==0.10.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==2.3.7
wheel==0.45.1
whois==1.20240129.2
XlsxWriter==3.2.3
yarl==1.20.0